FarmService and PointService reafactored. No interface required for services - @Autowired
