+++ src/main/java/com/warkoczewski/AgroAdventureBooking/service/FarmService.java
@@ -1,14 +1,26 @@
 package com.warkoczewski.AgroAdventureBooking.service;
 
 import com.warkoczewski.AgroAdventureBooking.model.Farm;
+import com.warkoczewski.AgroAdventureBooking.repository.FarmRepository;
 import org.springframework.stereotype.Service;
-
 import java.util.List;
 import java.util.Optional;
 @Service
-public interface FarmService {
+public class FarmService {
 
-    Optional<Farm> showFarmByName(String name);
+    private final FarmRepository farmRepository;
 
-    List<Farm> showFarmsByNamePhrase(String phrase);
+    public FarmService(FarmRepository farmRepository) {
+        this.farmRepository = farmRepository;
+    }
+
+
+    public Optional<Farm> showFarmByName(String name) {
+        return farmRepository.findFarmByName(name);
+    }
+
+
+    public List<Farm> showFarmsByNamePhrase(String phrase) {
+        return farmRepository.findFarmByNameIsContaining(phrase);
+    }
 }
