+++ src/main/java/com/warkoczewski/AgroAdventureBooking/service/PointService.java
@@ -1,11 +1,26 @@
 package com.warkoczewski.AgroAdventureBooking.service;
 
 import com.warkoczewski.AgroAdventureBooking.model.Point;
+import com.warkoczewski.AgroAdventureBooking.repository.PointRepository;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-public interface PointService {
 
-    public List<Point> showAll();
-    public List<Point> showFarmsByRegion(String region);
+@Service
+public class PointService {
+
+    private final PointRepository pointRepository;
+
+    public PointService(PointRepository pointRepository) {
+        this.pointRepository = pointRepository;
+    }
+
+    public List<Point> showAll(){
+        return pointRepository.findAll();
+    }
+
+    public List<Point> showFarmsByRegion(String region){
+        return pointRepository.findPointByRegion(region);
+    }
 }
